# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Gallery API",
        "description": "Internal HTTP API for the self-hosted media gallery — a **hybrid REST-RPC**\ninterface served under the `/api` base path. All responses are JSON.\n\n**Conventions**\n- Reads return `200` with the resource or collection.\n- Creates return `201` with the created resource.\n- Updates (`PUT`/`PATCH`) return `200` with the updated resource.\n- Deletes return `204` with no body.\n- Non-CRUD actions (login, migrate, scan, bulk operations) are `POST`\n  verbs that return `200` with a small result object.\n- Errors return `{ \"error\": \"<MachineCode>\", \"message\": \"<human text>\" }`\n  with a `4xx`/`5xx` status. `error` is a stable PascalCase code safe to\n  branch on; `message` is for display and may change.\n\n**Auth** — a Bearer token (`Authorization: Bearer <token>`) obtained from\n`POST /auth/login` (valid 24h). All `GET`s are public; state-changing\nmethods require a token, except a small public allowlist (media tagging\nand the batched `POST /media/by-ids` read).",
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Same-origin API base path"
        }
    ],
    "paths": {
...
        "/tag-categories": {
            "get": {
                "tags": [
                    "Tag Categories"
                ],
                "summary": "List tag categories",
                "description": "GET /tag-categories — All tag categories.",
                "operationId": "344bd2707a98746738eb6b0b3ed143ff",
                "responses": {
                    "200": {
                        "description": "All categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TagCategory"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tag Categories"
                ],
                "summary": "Create a tag category",
                "description": "POST /tag-categories — Create a tag category. Returns the created category.\nBody: { category_name, category_short, color?, description?, sort_order? }",
                "operationId": "dc114b35e6a6392f6c75717bf3f23651",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_name",
                                    "category_short"
                                ],
                                "properties": {
                                    "category_name": {
                                        "type": "string"
                                    },
                                    "category_short": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "sort_order": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The created category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagCategory"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tag-categories/{category_id}": {
            "put": {
                "tags": [
                    "Tag Categories"
                ],
                "summary": "Update a tag category",
                "description": "PUT /tag-categories/{category_id} — Update a tag category.\nReturns the updated category. Body: same fields as create.",
                "operationId": "2cda0d143e8dfc92355bb0452d037f19",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_name",
                                    "category_short"
                                ],
                                "properties": {
                                    "category_name": {
                                        "type": "string"
                                    },
                                    "category_short": {
                                        "type": "string"
                                    },
                                    "color": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "sort_order": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The updated category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagCategory"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CategoryNotFound",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tag Categories"
                ],
                "summary": "Delete a tag category",
                "description": "DELETE /tag-categories/{category_id} — Delete a category that has no tags.",
                "operationId": "c1c34eafc2deb4bbe13b3114487847b6",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "400": {
                        "description": "CategoryInUse",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CategoryNotFound",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "description": "Standard error envelope.",
                "properties": {
                    "error": {
                        "description": "Stable PascalCase machine code (e.g. MediaNotFound).",
                        "type": "string"
                    },
                    "message": {
                        "description": "Human-readable message for display.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "TagCategory": {
                "description": "A tag category (color, shortcode, sort order).",
                "properties": {
                    "category_id": {
                        "type": "integer"
                    },
                    "category_name": {
                        "type": "string"
                    },
                    "category_short": {
                        "description": "Short code (<= 5 chars) used in tag prefixes.",
                        "type": "string"
                    },
                    "color": {
                        "description": "Bulma/extended palette color name.",
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "sort_order": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Admin bearer token issued by POST /auth/login (valid 24 hours).",
                "scheme": "bearer"
            }
        }
    },
...
}